Move events to feature folder
