+++ src/features/Events/Components/eventsFilters.js
@@ -2,13 +2,13 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import styled, { keyframes } from 'styled-components'
 import { Flex, Box } from 'grid-styled'
-import { media } from '../../theme/media'
-import { Consumer } from '../../components/appContext'
-import iconClear from '../../theme/assets/images/icon-clear.svg'
-import iconClose from '../../theme/assets/images/icon-close.svg'
-import EventDateFilter from './filters/eventDateFilter'
-import EventFreeFilter from './filters/eventFreeFilter'
-import EventDropdownFilter from './filters/eventDropdownFilter'
+import { media } from '../../../theme/media'
+import { Consumer } from '../../../components/appContext'
+import iconClear from '../../../theme/assets/images/icon-clear.svg'
+import iconClose from '../../../theme/assets/images/icon-close.svg'
+import EventDateFilter from '../filters/eventDateFilter'
+import EventFreeFilter from '../filters/eventFreeFilter'
+import EventDropdownFilter from '../filters/eventDropdownFilter'
 
 const FilterWrapper = styled(Flex)`
   background-color: ${props => props.theme.colors.white};
