+++ src/features/Events/Components/eventsYouMayLike.js
@@ -2,10 +2,10 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 import moment from 'moment'
-import ChevronRight from '../../components/chevronRight'
-import { Consumer } from '../../components/appContext'
-import { Container, Row, Column } from '../../components/grid'
-import { media } from '../../theme/media'
+import ChevronRight from '../../../components/chevronRight'
+import { Consumer } from '../../../components/appContext'
+import { Container, Row, Column } from '../../../components/grid'
+import { media } from '../../../theme/media'
 import EventListingCard from './eventListingCard'
 
 const ViewAll = styled.a`
