+++ src/features/Events/index.js
@@ -1,0 +1,8 @@
+export { default as EventListingCard } from './Components/eventListingCard'
+export { default as EventTagList } from './Components/eventTagList'
+export { default as EventSchedule } from './Components/eventSchedule'
+export { default as EventsYouMayLike } from './Components/eventsYouMayLike'
+export { default as EventInfoCard } from './Components/eventInfoCard'
+export {
+  default as EventDirectionsSection,
+} from './Components/eventDirectionsSection'
