+++ src/pages/events.js
@@ -5,13 +5,13 @@ import moment from 'moment'
 import noScroll from 'no-scroll'
 import { media } from '../theme/media'
 import theme from '../theme/theme'
-import { EventListingCard } from '../components/events'
-import EventsFilters from '../components/events/eventsFilters'
+import { EventListingCard } from '../features/Events'
+import EventsFilters from '../features/Events/Components/eventsFilters'
 import ImageBanner from '../components/imageBanner'
 import Button from '../components/button'
 import { Container, Row, Column } from '../components/grid'
 import { Consumer } from '../components/appContext'
-import { filterByLimit } from '../components/events/helpers'
+import { filterByLimit } from '../features/Events/helpers'
 import { dateFormat } from '../constants'
 import filterIcon from '../theme/assets/images/icon-filters.svg'
